refactor: split lib.rs into focused modules
- provider.rs: BlockRef, LogFilter, Log, EthProvider, NetworkConfig
- abi.rs: ABI selectors, encode/decode helpers, formatting utils
- events.rs: Erc1056Event, parse/collect logic + collect_events tests
- network.rs: NetworkChain, InvalidNetwork, DecodedMethodSpecificId, parse_address_bytes
- vm.rs: VerificationMethod, VerificationMethodType, Pending* structs, decode_delegate_type
- resolver.rs: DIDEthr, apply_events, resolve_*, serialize_document + MockProvider tests
- lib.rs: module declarations + public re-exports + offline-only tests
